+++ tests/testthat/test-getTestStatus.R
@@ -1,6 +1,6 @@
 context("getTestStatus")
 
-test_that("canGetStatus", {
+test_that("canGetTestStatus", {
   WPT <- server()
   status <- getTestStatus(WPT, testId = "atestid", 
                            requestId = "areqid", dryRun = TRUE)  
@@ -10,10 +10,19 @@ test_that("canGetStatus", {
   )
 })
 
-test_that("canGetErrorFromGetStatus", {
+test_that("canGetErrorFromGetTestStatus", {
   WPT <- server()
   expect_error(
     getTestStatus(WPT, testId = "atestid", requestId = "areqid"), 
     "Server returned a statusCode = 400"   
   )
 })
+
+test_that("canGetTestStatusUsingCurrentTestId", {
+  WPT <- server()
+  history <- getHistory(WPT)
+  tID <- history[["Test ID"]][1]
+  status <- getTestStatus(WPT, testId = tID, requestId = "areqid")  
+  expect_identical(status$testId, tID)
+})
+