+++ NEWS.md
@@ -1,6 +1,7 @@
 # 0.0.4
 * Add getHistory function, tests and docs
 * Add option to use GET in sendQuery internal function
+* getTestStatus checks for statusCode <= 200L now
 
 # 0.0.3
 * Add getTesters function, tests and docs